extract regular expression and add Arabic and Thai support
