+++ lib/hashtagable.dart
@@ -1,6 +1,7 @@
 library hashtagable;
 
-export 'decorator.dart';
+export 'decorator/decorator.dart';
+export 'decorator/hashtag_regular_expression.dart';
 export 'functions.dart';
-export 'hashtag_text.dart';
-export 'hashtag_text_field.dart';
+export 'widgets/hashtag_text.dart';
+export 'widgets/hashtag_text_field.dart';
