+++ lib/widgets/hashtag_editable_text.dart
@@ -4,7 +4,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:hashtagable/decorator.dart';
+import 'package:hashtagable/decorator/decorator.dart';
 
 /// Show decorated tagged text while user is inputting text.
 ///
