+++ test/hashtagable_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:hashtagable/decorator.dart';
+import 'package:hashtagable/decorator/decorator.dart';
 import 'package:hashtagable/functions.dart';
 
 void main() {
